// src/app/interfaces/skilled.ts
export interface Skilled {
  id?: number;
  name: string;
  created_at?: any;
  updated_at?: any;
  motive_state?: any;
  availability?: any;
  modality?: string;
  autonomous?: any;
  contact_phone?: string;
  contact_email?: string;
  contact_country?: string;
  contact_linkedln?: any;
  conditions_percentage?: any;
  conditions_price_hour?: any;
  score?: any;
  nif?: any;
  email_credentials?: any;
  email_password_credentials?: any;
  zoom_credentials?: any;
  zoom_password_credentials?: any;
  photo_file?: any;
  cv_file?: any;
  observations?: any;
  origin?: any;
  state: string;
  tags: any[];
}